Add files via upload
------------------------------------------------------------------
Welcome to the Guide for Software Installation for Computer Vision
------------------------------------------------------------------

This guide will help you install a list of required software and libraries for Computer Vision and Deep Learning development on Ubuntu.

The name of the script is 'required_libraries.sh'.

Steps to Run the Script
-----------------------

1. Opening the Terminal:

   First, you will need to open the terminal. You can open the terminal by searching for 'terminal' in the system dash or by using the keyboard shortcut 'Ctrl+Alt+T'.

2. Navigating to the script:

   Once the terminal is open, you will need to navigate to the directory (folder) where the script 'required_libraries.sh' is saved.

   You can navigate in terminal using the 'cd' command which stands for 'change directory'. For example, if your script is in a folder on the Desktop called 'scripts', you would type:

   cd Desktop/scripts

   and press Enter. You'll now be in the scripts folder.

3. Running the Script:

   Before running the script, you need to make sure that it has the necessary permissions. In the terminal, type the following command:

   chmod +x required_libraries.sh

   and press Enter. This command changes the script to be executable, which means you can now run it.

   Now, to run the script, type:

   ./required_libraries.sh

   and press Enter. 

   The script will now begin running. Depending on your system's speed and the current state of installed software, it may take some time to complete. 

Please note that this script requires administrator (sudo) access, and will run commands as such. It is important to only run scripts as administrator if they are from a trusted source, as they have high-level access to your system.

Remember: You can stop the script at any time by pressing 'Ctrl+C' in the terminal.

In case of any error messages or if the script fails to run, make sure the command was typed correctly, and that the 'required_libraries.sh' script is indeed in the directory you navigated to.

If you are still facing issues, you may need to consult further documentation or support for your specific system configuration.